# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "NovaRewards API",
    "version": "1.0.0",
    "description": "REST API for the Nova Rewards blockchain loyalty platform. User endpoints require a Bearer JWT from `POST /auth/login`. Merchant endpoints require an `x-api-key` header.",
    "contact": {
      "name": "Nova Rewards Team"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Local development"
    },
    {
      "url": "https://api.novarewards.io/api",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token from `POST /auth/login`."
      },
      "merchantApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Merchant API key returned on registration (shown once)."
      }
    },
...
  },
  "paths": {
...
    "/trustline/verify": {
      "post": {
        "tags": [
          "Trustline"
        ],
        "summary": "Check whether a wallet has an active NOVA trustline",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "walletAddress"
                ],
                "properties": {
                  "walletAddress": {
                    "type": "string",
                    "example": "GBBD47..."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Trustline status",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "properties": {
                        "exists": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid wallet address"
          }
        }
      }
    },
    "/trustline/build-xdr": {
      "post": {
        "tags": [
          "Trustline"
        ],
        "summary": "Build unsigned changeTrust XDR for Freighter signing",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "walletAddress"
                ],
                "properties": {
                  "walletAddress": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "XDR envelope"
          },
          "400": {
            "description": "Invalid wallet address"
          }
        }
      }
    },
...
  }
}